+++ nn/src/activation.c
@@ -45,7 +45,7 @@ void nn_none_fn_d(Vector* z, Vector* d) {
 	if (z->dimension != d->dimension) {
 		fatal("Mismatched vector size, z: %zu d: %zu", z->dimension, d->dimension);
 	}
-	for (int i = 0; i < z->dimension; i++) {
+	for (size_t i = 0; i < z->dimension; i++) {
 		d->data[i] = 1.0f;
 	}
 }
@@ -58,7 +58,7 @@ void nn_relu(Vector* z, Vector* a) {
 	if (z->dimension != a->dimension) {
 		fatal("Mismatched vector size, z: %zu a: %zu", z->dimension, a->dimension);
 	}
-	for (int i = 0; i < z->dimension; i++) {
+	for (size_t i = 0; i < z->dimension; i++) {
 		if (z->data[i] > 0) a->data[i] = z->data[i];
 	}
 }
@@ -66,7 +66,7 @@ void nn_relu_d(Vector* z, Vector* d) {
 	if (z->dimension != d->dimension) {
 		fatal("Mismatched vector size, z: %zu d: %zu", z->dimension, d->dimension);
 	}
-	for (int i = 0; i < z->dimension; i++) {
+	for (size_t i = 0; i < z->dimension; i++) {
 		if (z->data[i] > 0) d->data[i] = 1.0f;
 	}
 }
@@ -80,7 +80,7 @@ void nn_sigmoid(Vector* z, Vector* a) {
 	if (z->dimension != a->dimension) {
 		fatal("Mismatched vector size, z: %zu a: %zu", z->dimension, a->dimension);
 	}
-	for (int i = 0; i < z->dimension; i++) {
+	for (size_t i = 0; i < z->dimension; i++) {
 		a->data[i] = _sigmoid(z->data[i]);
 	}
 }
@@ -88,7 +88,7 @@ void nn_sigmoid_d(Vector* z, Vector* d) {
 	if (z->dimension != d->dimension) {
 		fatal("Mismatched vector size, z: %zu d: %zu", z->dimension, d->dimension);
 	}
-	for (int i = 0; i < z->dimension; i++) {
+	for (size_t i = 0; i < z->dimension; i++) {
 		float sig = _sigmoid(z->data[i]);
 		d->data[i] = sig*(1-sig);
 	}
@@ -104,17 +104,17 @@ void nn_softmax(Vector* z, Vector* a) {
 	}
 	float exp_sum = 0;
 	float z_max = 0;
-	for (int i = 0; i < z->dimension; i++) {
+	for (size_t i = 0; i < z->dimension; i++) {
 		if (z->data[i] > z_max) {
 			z_max = z->data[i];
 		}
 	}
-	for (int i = 0; i < z->dimension; i++) {
+	for (size_t i = 0; i < z->dimension; i++) {
 		float z_exp = exp(z->data[i] - z_max);
 		exp_sum += exp(z->data[i] - z_max);
 		a->data[i] = z_exp;
 	}
-	for (int i = 0; i < z->dimension; i++) {
+	for (size_t i = 0; i < z->dimension; i++) {
 		a->data[i] /= exp_sum;
 	}
 }
